+++ src/main/java/com/jangburich/application/order/service/OrderQueryService.java
@@ -3,7 +3,7 @@ package com.jangburich.application.order.service;
 import com.jangburich.domain.entity.Orders;
 import com.jangburich.domain.repository.OrdersRepository;
 import com.jangburich.domain.user.domain.User;
-import com.jangburich.domain.user.repository.UserRepository;
+import com.jangburich.domain.repository.UserRepository;
 import com.jangburich.global.error.DefaultException;
 import com.jangburich.global.payload.ErrorCode;
 import com.jangburich.presentation.order.dto.response.OrderResponse;
