+++ src/main/java/com/jangburich/application/store/service/io/StoreFileService.java
@@ -37,7 +37,7 @@ import com.jangburich.presentation.store.dtos.response.store.StoreTeamResponseDT
 import com.jangburich.domain.repository.StoreRepository;
 import com.jangburich.domain.repository.StoreTeamRepository;
 import com.jangburich.domain.user.domain.User;
-import com.jangburich.domain.user.repository.UserRepository;
+import com.jangburich.domain.repository.UserRepository;
 import com.jangburich.global.error.DefaultNullPointerException;
 import com.jangburich.global.payload.ErrorCode;
