+++ src/main/java/com/jangburich/domain/owner/domain/service/OwnerCommandService.java
@@ -7,7 +7,7 @@ import com.jangburich.domain.owner.domain.controller.dto.req.OwnerCreateReqDTO;
 import com.jangburich.domain.owner.domain.entity.Owner;
 import com.jangburich.domain.owner.domain.repository.OwnerRepository;
 import com.jangburich.domain.user.domain.User;
-import com.jangburich.domain.user.repository.UserRepository;
+import com.jangburich.domain.repository.UserRepository;
 import com.jangburich.global.error.DefaultNullPointerException;
 import com.jangburich.global.payload.ErrorCode;
