+++ src/main/java/com/jangburich/domain/owner/domain/service/OwnerQueryService.java
@@ -11,7 +11,7 @@ import com.jangburich.domain.owner.domain.controller.dto.res.OwnerResDto;
 import com.jangburich.domain.owner.domain.entity.Owner;
 import com.jangburich.domain.owner.domain.repository.OwnerRepository;
 import com.jangburich.domain.user.domain.User;
-import com.jangburich.domain.user.repository.UserRepository;
+import com.jangburich.domain.repository.UserRepository;
 import com.jangburich.global.error.DefaultNullPointerException;
 import com.jangburich.global.payload.ErrorCode;
