+++ src/main/java/com/jangburich/domain/repository/queryDsl/UserQueryDslRepositoryImpl.java
@@ -1,4 +1,4 @@
-package com.jangburich.domain.user.repository;
+package com.jangburich.domain.repository.queryDsl;
 
 import static com.jangburich.domain.entity.QStoreTeam.storeTeam;
 import static com.jangburich.domain.team.domain.QTeam.*;
@@ -6,6 +6,7 @@ import static com.jangburich.domain.team.domain.QUserTeam.userTeam;
 import static com.jangburich.domain.user.domain.QUser.user;
 
 import com.jangburich.domain.common.Status;
+import com.jangburich.domain.repository.queryDsl.UserQueryDslRepository;
 import com.jangburich.domain.user.dto.response.QTeamsResponse;
 import com.jangburich.domain.user.dto.response.QUserHomeResponse;
 import com.jangburich.domain.user.dto.response.TeamsResponse;
