+++ src/main/java/com/jangburich/domain/team/application/TeamService.java
@@ -21,7 +21,7 @@ import com.jangburich.domain.team.domain.repository.UserTeamRepository;
 import com.jangburich.domain.team.dto.request.RegisterTeamRequest;
 import com.jangburich.domain.team.dto.response.MyTeamDetailResponse;
 import com.jangburich.domain.user.domain.User;
-import com.jangburich.domain.user.repository.UserRepository;
+import com.jangburich.domain.repository.UserRepository;
 import com.jangburich.global.payload.Message;
 
 import lombok.RequiredArgsConstructor;
